Migrate IncentiveStore into Donation and EventDetails stores
The `incentives` module will go away soon, since it's really just supporting EventDetails and Donation already.
