+++ bundles/donate/Store.ts
@@ -4,10 +4,8 @@ import thunk from 'redux-thunk';
 import { Action } from './Action';
 import DonationReducer from './donation/DonationReducer';
 import EventDetailsReducer from './event_details/EventDetailsReducer';
-import IncentivesReducer from './incentives/IncentivesReducer';
 
 const combinedReducer = combineReducers({
-  incentives: IncentivesReducer,
   eventDetails: EventDetailsReducer,
   donation: DonationReducer,
 });
