+++ bundles/donate/donation/DonationStore.ts
@@ -1,8 +1,8 @@
+import _ from 'lodash';
 import { createSelector } from 'reselect';
 
 import * as CurrencyUtils from '../../public/util/currency';
 import * as EventDetailsStore from '../event_details/EventDetailsStore';
-import * as IncentiveStore from '../incentives/IncentiveStore';
 import { StoreState } from '../Store';
 
 type DonationValidation = {
@@ -10,7 +10,8 @@ type DonationValidation = {
   errors: Array<{ field: string; message: string }>;
 };
 
-const getDonationState = (state: StoreState) => state.donation;
+const getDonationState = (state: StoreState) => state.donation.donation;
+const getBidsById = (state: StoreState) => state.donation.bids;
 
 export const getDonation = getDonationState;
 
@@ -19,8 +20,18 @@ export const getDonationAmount = createSelector(
   donation => donation.amount,
 );
 
+export const getBids = createSelector(
+  [getBidsById],
+  bidsById => Object.values(bidsById),
+);
+
+export const getAllocatedBidTotal = createSelector(
+  [getBids],
+  bids => _.sumBy(bids, 'amount'),
+);
+
 export const validateDonation = createSelector(
-  [getDonationState, IncentiveStore.getBids, EventDetailsStore.getEventDetails],
+  [getDonationState, getBids, EventDetailsStore.getEventDetails],
   (donation, bids, eventDetails): DonationValidation => {
     const validation: DonationValidation = { valid: true, errors: [] };
