--- bundles/public/util/currency.js
@@ -1,3 +1,0 @@
-export function asCurrency(amount) {
-  return `$${Number(amount).toFixed(2)}`;
-}
